refactor: extract buildFilter/buildSort to testable module and document multi-currency behavior
